+++ react/components/ShippingSimulator/Wrapper.js
@@ -49,6 +49,8 @@ const BaseShippingSimulatorWrapper = ({
   onShippingAddressUpdate = _ => {},
   initialPostalCode = undefined,
   shouldUpdateOrderForm,
+  pricingMode,
+  selectedQuantity,
 }) => {
   const [shipping, setShipping] = useState(null)
   const [loading, setLoading] = useState(false)
@@ -74,7 +76,7 @@ const BaseShippingSimulatorWrapper = ({
             postalCode: rawAddress.postalCode,
             items: [
               {
-                quantity: '1',
+                quantity: selectedQuantity,
                 id: skuId,
                 seller,
               },
@@ -102,6 +104,7 @@ const BaseShippingSimulatorWrapper = ({
       country,
       seller,
       skuId,
+      selectedQuantity,
       onShippingAddressUpdate,
       shouldUpdateOrderForm,
     ]
@@ -133,6 +136,8 @@ const BaseShippingSimulatorWrapper = ({
       address={address}
       isValid={isValid}
       shipping={shipping}
+      pricingMode={pricingMode}
+      selectedQuantity={selectedQuantity}
       onAddressChange={updateAddress}
       onCalculateShipping={handleCalculateShipping}
     />
@@ -145,6 +150,8 @@ const ShippingSimulatorWithOrderForm = ({
   seller,
   loaderStyles,
   shouldUpdateOrderForm,
+  pricingMode,
+  selectedQuantity,
 }) => {
   const { updateSelectedAddress } = useOrderShipping()
   const { orderForm } = useOrderForm()
@@ -164,6 +171,8 @@ const ShippingSimulatorWithOrderForm = ({
       }
       onShippingAddressUpdate={updateSelectedAddress}
       shouldUpdateOrderForm={shouldUpdateOrderForm}
+      pricingMode={pricingMode}
+      selectedQuantity={selectedQuantity}
     />
   )
 }
@@ -184,6 +193,9 @@ const ShippingSimulatorWrapper = props => {
 
   const country = props.country || culture.country
   const skuId = props.skuId || productContext?.selectedItem?.itemId
+  const selectedQuantity = productContext?.selectedQuantity?.toString()
+
+  const { pricingMode } = props
 
   let sellerId = props.seller
 
@@ -208,6 +220,8 @@ const ShippingSimulatorWrapper = props => {
         country={country}
         skuId={skuId}
         seller={sellerId}
+        pricingMode={pricingMode}
+        selectedQuantity={selectedQuantity}
         loaderStyles={props.loaderStyles}
       />
     )
@@ -221,6 +235,8 @@ const ShippingSimulatorWrapper = props => {
           seller={sellerId}
           skuId={skuId}
           loaderStyles={props.loaderStyles}
+          pricingMode={pricingMode}
+          selectedQuantity={selectedQuantity}
           shouldUpdateOrderForm={shouldUpdateOrderForm}
         />
       </OrderFormLoader>